Move HLSL and DirectX/SPIRV support to the trunk build
When @llvm-beanz attempted to add HLSL support here he added this
under the `llvm-*` build targets, but those are actually used for the
"library" builds of LLVM, not the compilers themselves.

Move this support under the trunk part of the `*` target like other
experimental targets. This should make it so that the current "clang"
option for HLSL stops giving errors about missing "hlsl.h" and should
be useable for basic compute shaders.