Move HLSL and DirectX/SPIRV support to the trunk build (Attempt 2)
In my previous attempt at doing this in 1e799cd I misread the case
labels and actually modified the build for clang with a specific
version, rather than modifying trunk like I said I was doing.

Here we revert the "specific-version that is above clang-12" case to
what it was before and actually put the DirectX, SPIRV, and HLSL flags
in the trunk build itself.