Add support for ClangIR llvm fork
ClangIR is a new IR for clang (clangir.org) hosted in LLVM's umbrella
of projects. ClangIR is a C/C++ dialect for MLIR built into Clang.

Enable support for ClangIR as part of the compiler-explorer by setting
up repos and build configurations.

I'm a big fan of the compiler-explorer and would be great if we can
incorporate support for our project, thanks! :)

-